+++ src/widget/toggle_button_widget.rs
@@ -43,7 +43,6 @@ impl ToggleButtonWidget {
         text: String,
         font_size: u32,
         justify: TextJustify,
-        selected: bool,
     ) -> Self {
         Self {
             config: Configurable::new(),
@@ -55,7 +54,7 @@ impl ToggleButtonWidget {
                 font_size,
                 justify,
             ),
-            selected,
+            selected: false,
             active: false,
         }
     }
@@ -97,6 +96,13 @@ impl Widget for ToggleButtonWidget {
         self.config().set(config, config_value.clone());
         self.base_widget.set_config(config, config_value.clone());
         self.text_widget.set_config(config, config_value.clone());
+
+        if config == CONFIG_SELECTED {
+            if self.config().get_toggle(CONFIG_SELECTED) {
+                self.draw_hovered();
+                self.selected = true;
+            }
+        }
     }
 
     fn handle_event(&mut self, injected: bool, event: CallbackEvent) -> Option<CallbackEvent> {